plot: add exploded scatter plots, violin plot and ridge line plot tweaks
- ridgelines: better colors, jittered points, sizes
- violins: scale adjustments, jittered points, sizes
- exploded scatter plots, global and per class

exploded scatter plots rely on geom_density_2d, which in turn uses
MASS::bandwidth.nrd to estimate bandwith; in some cases this calculation can
fail and it will yield an empty plot.
a fix is to compute bandwidth ourselves and set it to something like 0.1 if
it's 0, then pass it to stat_density_2d.  but, woah there.